+++ ploy/tests/dummy_plugin.py
@@ -42,7 +42,7 @@ class Instance(BaseInstance, StartupScriptMixin):
         self.startup_script(overrides=overrides)
         log.info('start: %s %s', self.id, overrides)
         # this is here to get full coverage of the cmd_start method in common.py
-        if overrides.keys() != ['servers']:
+        if overrides.keys() != ['instances']:
             return overrides
 
     def status(self):